both collect render and farm in same file
