+++ client/ayon_harmony/plugins/publish/collect_render.py
@@ -1,7 +1,21 @@
 # -*- coding: utf-8 -*-
-"""Collect data for render farm renders from scene."""
+"""Collect render instances in Harmony."""
+
+import pyblish.api
 
 from ayon_harmony.api.plugins import publish
+
+
+class CollectRender(publish.CollectHarmonyRender):
+    """Collect render instances from Harmony.
+
+    Create regular render instances based on the ones created within the publisher.
+    """
+
+    label = "Collect Render Instances"
+    order = pyblish.api.CollectorOrder + 0.01
+    hosts = ["harmony"]
+    families = ["render"]
 
 
 class CollectFarmRender(publish.CollectHarmonyRender):
